Added pages About, Mentoring, Consulting, Academies, Contact Us, Free Consultation. Made some styling changes to home page.

+++ backend/src/main/java/com/shifterwebapp/shifter/enrollment/service/EnrollmentService.java
@@ -11,20 +11,22 @@ import com.shifterwebapp.shifter.enrollment.EnrollmentRepository;
 import com.shifterwebapp.shifter.enums.*;
 import com.shifterwebapp.shifter.exception.AlreadyEnrolledException;
 import com.shifterwebapp.shifter.exception.PaymentNotCompleteException;
+import com.shifterwebapp.shifter.external.email.EmailService;
 import com.shifterwebapp.shifter.payment.Payment;
 import com.shifterwebapp.shifter.payment.service.PaymentService;
 import com.shifterwebapp.shifter.user.service.UserService;
 import com.shifterwebapp.shifter.usercourseprogress.UserCourseProgress;
 import com.shifterwebapp.shifter.usercourseprogress.UserCourseProgressRepository;
+import com.shifterwebapp.shifter.usercourseprogress.service.UserCourseProgressService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.util.Date;
+import java.time.LocalDate;
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class EnrollmentService implements ImplEnrollmentService{
+public class EnrollmentService implements ImplEnrollmentService {
 
     private final EnrollmentRepository enrollmentRepository;
     private final CourseRepository courseRepository;
@@ -32,6 +34,7 @@ public class EnrollmentService implements ImplEnrollmentService{
     private final UserService userService;
     private final PaymentService paymentService;
     private final EnrollmentMapper enrollmentMapper;
+    private final EmailService emailService;
     private final Validate validate;
 
     @Override
@@ -90,7 +93,7 @@ public class EnrollmentService implements ImplEnrollmentService{
 
         Enrollment enrollment = Enrollment.builder()
                 .enrollmentStatus(EnrollmentStatus.PENDING)
-                .date(new Date())
+                .date(LocalDate.now())
                 .payment(payment)
                 .review(null)
                 .course(course)
@@ -113,6 +116,20 @@ public class EnrollmentService implements ImplEnrollmentService{
 
         userCourseProgressRepository.saveAll(progressList);
 
+        String courseTitleFormatted = course.getTitleShort()
+                .toLowerCase()
+                .trim()
+                .replaceAll("\\s+", "-")
+                .replaceAll("-+", "-");
+
+        emailService.sendCoursePurchaseConfirmation(
+                payment.getUser().getEmail(),
+                course.getTitle(),
+                course.getDescription(),
+                "http://localhost:5173/learn/" + course.getId() + "/" + courseTitleFormatted,
+                payment.getUser().getName()
+        );
+
         return enrollmentMapper.toDto(enrollment);
     }
 
@@ -125,32 +142,45 @@ public class EnrollmentService implements ImplEnrollmentService{
     }
 
     @Override
-    public EnrollmentDto updateEnrollmentStatusToActive(Long enrollmentId) {
-        validate.validateEnrollmentExists(enrollmentId);
+    public EnrollmentDto updateEnrollmentStatusToActive(Enrollment enrollment) {
+        validate.validateEnrollmentExists(enrollment.getId());
 
-        Enrollment enrollment = enrollmentRepository.findById(enrollmentId).orElseThrow();
-        enrollment.setEnrollmentStatus(EnrollmentStatus.ACTIVE);
-        enrollmentRepository.save(enrollment);
+        if (enrollment.getEnrollmentStatus() == EnrollmentStatus.PENDING) {
+            enrollment.setEnrollmentStatus(EnrollmentStatus.ACTIVE);
+            enrollment.setActivatedAt(LocalDate.now());
+            enrollmentRepository.save(enrollment);
+        }
 
         return enrollmentMapper.toDto(enrollment);
     }
 
-    // CALLING user SERVICE HERE. IS THERE A BETTER WAY FOR THIS ???
     @Override
     public EnrollmentDto updateEnrollmentStatusToCompleted(Long enrollmentId) {
         validate.validateEnrollmentExists(enrollmentId);
 
+        List<UserCourseProgress> userCourseProgresses = userCourseProgressRepository.findByEnrollmentId(enrollmentId);
+
+        boolean allCompleted = userCourseProgresses.stream().allMatch(UserCourseProgress::isCompleted);
+
         Enrollment enrollment = enrollmentRepository.findById(enrollmentId).orElseThrow();
-        enrollment.setEnrollmentStatus(EnrollmentStatus.COMPLETED);
 
-        Long userId = enrollment.getPayment().getUser().getId();
-        List<String> skillsGained = enrollment.getCourse().getSkillsGained();
-        userService.addPoints(userId, PointsConstants.BUY_COURSE);
-        userService.addSkills(userId, skillsGained);
-        userService.removeDesiredSkills(userId, skillsGained);
+        if (allCompleted) {
+            enrollment.setEnrollmentStatus(EnrollmentStatus.COMPLETED);
+            enrollment.setCompletedAt(LocalDate.now());
+            enrollmentRepository.save(enrollment);
 
-        enrollmentRepository.save(enrollment);
+            Long userId = enrollment.getPayment().getUser().getId();
+            List<String> skillsGained = enrollment.getCourse().getSkillsGained();
+            userService.addPoints(userId, PointsConstants.BUY_COURSE);
+            userService.addSkills(userId, skillsGained);
+            userService.removeDesiredSkills(userId, skillsGained);
+        }
 
         return enrollmentMapper.toDto(enrollment);
+    }
+
+    @Override
+    public Enrollment saveEnrollment(Enrollment enrollment) {
+        return enrollmentRepository.save(enrollment);
     }
 }
